Dir enum has been moved to own header #403
