+++ sources/core/ChatFrameInfo.h
@@ -3,7 +3,7 @@
 #include <QString>
 #include <QHash>
 
-#include "core_headers/KvGlobals.h"
+#include "core/KvGlobals.h"
 
 #include "Hearer.h"
