+++ sources/core/Hearer.h
@@ -2,7 +2,7 @@
 
 #include <QVector>
 
-#include "core_headers/KvGlobals.h"
+#include "core/KvGlobals.h"
 
 namespace kv
 {
