+++ sources/core/objects/Object.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "core_headers/KvAbort.h"
-#include "core_headers/KvGlobals.h"
+#include "core/KvGlobals.h"
 #include "core_headers/Hashes.h"
 
 #include "../FastIsType.h"
