+++ sources/representation/GLSprite.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 
-#include "core_headers/KvGlobals.h"
+#include "core/KvGlobals.h"
 
 #include "GLSprite.h"
